Regarding Warning and ARC fixes

+++ PDFKitten/CIDFont.m
@@ -5,7 +5,7 @@
 - (NSString *)stringWithPDFString:(CGPDFStringRef)pdfString
 {
 	unichar *characterIDs = (unichar *) CGPDFStringGetBytePtr(pdfString);
-	int length = CGPDFStringGetLength(pdfString) / sizeof(unichar);
+	int length = (int)(CGPDFStringGetLength(pdfString) / sizeof(unichar));
 	int magicalOffset = ([self isIdentity] ? 0 : 30);
 	NSMutableString *unicodeString = [NSMutableString string];
 	for (int i = 0; i < length; i++)
